// extension/src/types/storage.ts
export type LocaleStorageKey =
  | "localePhrases"
  | "localeLang"
  | "localeHash"
  | "localeVersion"
  | "localeUpdatedAt"
  | "localeLangOverride";